backend/requirements.txt
Flask==3.1.*
Flask-SQLAlchemy==3.1.*
Flask-Migrate==4.0.*
Flask-JWT-Extended==4.7.*
Flask-CORS==5.0.*
marshmallow>=3.18,<4
flask-marshmallow==1.2.*
marshmallow-sqlalchemy==1.1.*
psycopg2-binary==2.9.*
APScheduler==3.10.*
requests==2.32.*
python-dotenv==1.0.*
gunicorn==23.*
pytest==8.*
pytest-flask==1.*
